requirements.txt
dj-database-url==2.0.0
Django==4.2.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.11
python-dotenv==1.0.0
Pillow==9.5.0